pyproject.toml
[tool.poetry]
name = "py-qiita-search"
version = "0.2.0"
description = ""
authors = ["Yaboxi <z.yaboxi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
tqdm = "^4.65.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
